Fix db password parsing in crate (#735)

+++ src/utils/database-common/src/db_provider.rs
@@ -21,7 +21,7 @@ pub enum DatabaseProvider {
 }
 
 impl DatabaseProvider {
-    pub fn default_port(&self) -> u32 {
+    pub fn default_port(&self) -> u16 {
         match self {
             DatabaseProvider::MariaDB | DatabaseProvider::MySql => 3306,
             DatabaseProvider::Postgres => 5432,
